pyproject.toml
[tool.poetry]
name = "lumenpay-bot"
version = "0.1.0"
description = "LumenPay Bot – платформа для управления платными подписками в Telegram."
authors = ["Lumen Team <team@example.com>"]
license = "Proprietary"
readme = "README.md"
packages = [
  { include = "backend" },
  { include = "bot" },
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
aiosqlite = "^0.20.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
bcrypt = "<4.0.0"
python-dateutil = "^2.9.0"
pytz = "^2024.1"
httpx = "^0.27.0"
apscheduler = "^3.10.4"
redis = "^5.0.7"
cryptography = "^43.0.0"
slowapi = "^0.1.9"
structlog = "^24.1.0"
loguru = "^0.7.3"
orjson = "^3.10.12"
tenacity = "^8.3.0"
yookassa = "^3.5.0"
aioboto3 = "^12.3.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
python-multipart = "^0.0.9"
python-telegram-bot = "^21.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.8"
pytest-cov = "^5.0.0"
httpx = { version = "^0.27.0", extras = ["cli"] }
mypy = "^1.10.0"
ruff = "^0.6.1"
black = "^24.8.0"
isort = "^5.13.2"
types-python-dateutil = "^2.9.0.20240316"
types-redis = "^4.6.0.20240418"
types-requests = "^2.32.0.20240622"

[tool.poetry.scripts]
lumenpay-backend = "backend.app.main:run"
lumenpay-telegram-bot = "bot.app.main:run"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
combine_as_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
  "E",
  "F",
  "I",
  "N",
  "UP",
  "ASYNC",
  "S",
  "B",
  "A",
  "C4",
  "DTZ",
  "PT",
  "T20",
  "Q",
  "RUF",
]
ignore = [
  "B008",  # do not perform function call in argument defaults
  "DTZ005",  # datetime.now tz aware handled manually
  "RUF001",  # allow Cyrillic text in user-facing messages
]

[tool.ruff.lint.per-file-ignores]
"backend/alembic/env.py" = ["E402"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --disable-warnings --strict-markers"
testpaths = [
  "backend/tests",
]

[tool.coverage.run]
branch = true
source = [
  "backend/app",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 80

